refactor(CommentService): 将 CommentService 接口中的参数类型从 Long 改为 Integer
- 修改 getCommentsByPostId 方法的参数类型从 Long 改为 Integer
- 修改 deleteCommentById 方法的参数类型从 Long 改为 Integer

+++ backend/src/main/java/cn/hmck/service/CommentService.java
@@ -35,7 +35,7 @@ public interface CommentService extends IService<Comment> {
      * @param postId 动态ID
      * @return 评论列表
      */
-    List<Comment> getCommentsByPostId(Long postId);
+    List<Comment> getCommentsByPostId(Integer postId);
 
     /**
      * 删除评论或取消点赞
@@ -43,5 +43,5 @@ public interface CommentService extends IService<Comment> {
      * @param id 评论或点赞的ID
      * @return 是否成功
      */
-    boolean deleteCommentById(Long id);
+    boolean deleteCommentById(Integer id);
 }
